+++ examples/hackernews/astro.config.mjs
@@ -1,6 +1,7 @@
 // @ts-check
-import { defineConfig } from 'astro/config';
+
 import node from '@astrojs/node';
+import { defineConfig } from 'astro/config';
 
 // https://astro.build/config
 export default defineConfig({