+++ packages/astro/e2e/view-transitions.test.js
@@ -878,8 +878,8 @@ test.describe('View Transitions', () => {
 		// go to external page
 		await page.click('#click-redirect-external');
 		// doesn't work for playwright when we are too fast
-
-		await page.waitForURL('http://example.com');
+		await page.waitForLoadState('commit', { timeout: 5000 });
+		await page.waitForURL('http://example.com/', { waitUntil: 'commit', timeout: 5000 });
 		await expect(page.locator('h1'), 'should have content').toHaveText('Example Domain');
 		expect(loads.length, 'There should be 2 page loads').toEqual(2);
 	});